feat: cozy house, clear visa 신청 폼 API 구현
