+++ src/main/java/site/cleanfree/be_main/common/accessip/AccessIpController.java
@@ -14,11 +14,11 @@ import site.cleanfree.be_main.common.ClientIpAccessor;
 import site.cleanfree.be_main.common.exception.ErrorStatus;
 import site.cleanfree.be_main.consultant.ConsultantService;
 import site.cleanfree.be_main.cookingstation.CookingStationService;
-import site.cleanfree.be_main.cozyhouse.CozyHouseService;
+import site.cleanfree.be_main.cozyhouse.application.CozyHouseService;
 import site.cleanfree.be_main.cozyquick.application.CozyquickService;
 import site.cleanfree.be_main.createvalue.application.CreatevalueService;
 import site.cleanfree.be_main.curesilver.CureSilverService;
-import site.cleanfree.be_main.visa.VisaService;
+import site.cleanfree.be_main.visa.application.VisaService;
 
 @Slf4j
 @RestController
