+++ src/main/java/site/cleanfree/be_main/cozyhouse/application/CozyHouseService.java
@@ -1,11 +1,15 @@
-package site.cleanfree.be_main.cozyhouse;
+package site.cleanfree.be_main.cozyhouse.application;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import site.cleanfree.be_main.common.BaseResponse;
 import site.cleanfree.be_main.common.exception.ErrorStatus;
-import site.cleanfree.be_main.visa.VisaAccess;
+import site.cleanfree.be_main.cozyhouse.domain.CozyHouse;
+import site.cleanfree.be_main.cozyhouse.domain.CozyHouseAccess;
+import site.cleanfree.be_main.cozyhouse.infrastructure.CozyHouseAccessRepository;
+import site.cleanfree.be_main.cozyhouse.infrastructure.CozyHouseRepository;
+import site.cleanfree.be_main.cozyhouse.vo.CozyHouseRegisterRequestVo;
 
 @Service
 @RequiredArgsConstructor
@@ -14,13 +18,16 @@ public class CozyHouseService {
     private final CozyHouseRepository cozyHouseRepository;
     private final CozyHouseAccessRepository cozyHouseAccessRepository;
 
-    public BaseResponse<Object> register(String clientIp) {
+    public BaseResponse<Object> register(String clientIp,
+        CozyHouseRegisterRequestVo cozyHouseRegisterRequestVo) {
         Optional<CozyHouse> cozyHouseOpt = cozyHouseRepository.findCozyHouseByIp(clientIp);
 
         if (cozyHouseOpt.isEmpty()) {
             try {
                 cozyHouseRepository.save(CozyHouse.builder()
                     .ip(clientIp)
+                    .name(cozyHouseRegisterRequestVo.name())
+                    .phoneNumber(cozyHouseRegisterRequestVo.phoneNumber())
                     .build());
                 return BaseResponse.builder()
                     .success(true)
@@ -47,7 +54,8 @@ public class CozyHouseService {
     }
 
     public BaseResponse<Object> access(String clientIp) {
-        Optional<CozyHouseAccess> cozyHouseAccessOpt = cozyHouseAccessRepository.findCozyHouseAccessByIp(clientIp);
+        Optional<CozyHouseAccess> cozyHouseAccessOpt = cozyHouseAccessRepository.findCozyHouseAccessByIp(
+            clientIp);
 
         if (cozyHouseAccessOpt.isPresent()) {
             CozyHouseAccess cozyHouseAccess = cozyHouseAccessOpt.get();
