+++ src/main/java/site/cleanfree/be_main/cozyhouse/domain/CozyHouseAccess.java
@@ -1,4 +1,4 @@
-package site.cleanfree.be_main.cozyhouse;
+package site.cleanfree.be_main.cozyhouse.domain;
 
 import jakarta.persistence.Id;
 import lombok.Builder;
@@ -8,17 +8,19 @@ import org.springframework.data.mongodb.core.mapping.Document;
 import site.cleanfree.be_main.common.MongoBaseTimeEntity;
 
 @Getter
-@Document(collection = "cozy_house")
-public class CozyHouse extends MongoBaseTimeEntity {
+@Document(collection = "cozy_house_access")
+public class CozyHouseAccess extends MongoBaseTimeEntity {
 
     @Id
     private String id;
     @Indexed(unique = true)
     private String ip;
+    private int count;
 
     @Builder
-    public CozyHouse(String id, String ip) {
+    public CozyHouseAccess(String id, String ip, int count) {
         this.id = id;
         this.ip = ip;
+        this.count = count;
     }
 }
