+++ src/main/java/site/cleanfree/be_main/cozyhouse/infrastructure/CozyHouseRepository.java
@@ -1,7 +1,8 @@
-package site.cleanfree.be_main.cozyhouse;
+package site.cleanfree.be_main.cozyhouse.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import site.cleanfree.be_main.cozyhouse.domain.CozyHouse;
 
 public interface CozyHouseRepository extends MongoRepository<CozyHouse, String> {
