+++ src/main/java/site/cleanfree/be_main/cozyhouse/presentation/CozyHouseController.java
@@ -1,4 +1,4 @@
-package site.cleanfree.be_main.cozyhouse;
+package site.cleanfree.be_main.cozyhouse.presentation;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.cleanfree.be_main.common.BaseResponse;
 import site.cleanfree.be_main.common.ClientIpAccessor;
-import site.cleanfree.be_main.visa.VisaRegisterRequestVo;
+import site.cleanfree.be_main.cozyhouse.application.CozyHouseService;
+import site.cleanfree.be_main.cozyhouse.vo.CozyHouseRegisterRequestVo;
 
 @Slf4j
 @RestController
@@ -23,9 +24,9 @@ public class CozyHouseController {
     @PostMapping("/register")
     public ResponseEntity<BaseResponse<Object>> register(
         HttpServletRequest request,
-        @RequestBody VisaRegisterRequestVo visaRegisterRequestVo
+        @RequestBody CozyHouseRegisterRequestVo cozyHouseRegisterRequestVo
     ) {
         String clientIp = ClientIpAccessor.getIp(request);
-        return ResponseEntity.ok(cozyHouseService.register(clientIp));
+        return ResponseEntity.ok(cozyHouseService.register(clientIp, cozyHouseRegisterRequestVo));
     }
 }
