+++ src/main/java/site/cleanfree/be_main/cozyhouse/vo/CozyHouseRegisterRequestVo.java
@@ -1,0 +1,6 @@
+package site.cleanfree.be_main.cozyhouse.vo;
+
+
+public record CozyHouseRegisterRequestVo(String name, String phoneNumber) {
+
+}
