--- src/main/java/site/cleanfree/be_main/visa/VisaRegisterRequestVo.java
@@ -1,9 +1,0 @@
-package site.cleanfree.be_main.visa;
-
-import lombok.Getter;
-
-@Getter
-public class VisaRegisterRequestVo {
-
-    private String ip;
-}
