+++ src/main/java/site/cleanfree/be_main/visa/application/VisaService.java
@@ -1,11 +1,15 @@
-package site.cleanfree.be_main.visa;
+package site.cleanfree.be_main.visa.application;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import site.cleanfree.be_main.common.BaseResponse;
-import site.cleanfree.be_main.common.accessip.IpSaveRequestVo;
 import site.cleanfree.be_main.common.exception.ErrorStatus;
+import site.cleanfree.be_main.visa.domain.Visa;
+import site.cleanfree.be_main.visa.domain.VisaAccess;
+import site.cleanfree.be_main.visa.infrastructure.VisaAccessRepository;
+import site.cleanfree.be_main.visa.infrastructure.VisaRepository;
+import site.cleanfree.be_main.visa.vo.VisaRegisterRequestVo;
 
 @Service
 @RequiredArgsConstructor
@@ -14,7 +18,8 @@ public class VisaService {
     private final VisaRepository visaRepository;
     private final VisaAccessRepository visaAccessRepository;
 
-    public BaseResponse<Object> register(String clientIp) {
+    public BaseResponse<Object> register(String clientIp,
+        VisaRegisterRequestVo visaRegisterRequestVo) {
         Optional<Visa> visaOpt = visaRepository.findVisaByIp(clientIp);
 
         if (visaOpt.isPresent()) {
@@ -27,6 +32,8 @@ public class VisaService {
 
         visaRepository.save(Visa.builder()
             .ip(clientIp)
+            .name(visaRegisterRequestVo.name())
+            .phoneNumber(visaRegisterRequestVo.phoneNumber())
             .build());
 
         return BaseResponse.builder()
@@ -43,9 +50,9 @@ public class VisaService {
         if (visaAccessOpt.isPresent()) {
             VisaAccess visaAccess = visaAccessOpt.get();
             visaAccessRepository.save(VisaAccess.builder()
-                    .id(visaAccess.getId())
-                    .ip(visaAccess.getIp())
-                    .count(visaAccess.getCount() + 1)
+                .id(visaAccess.getId())
+                .ip(visaAccess.getIp())
+                .count(visaAccess.getCount() + 1)
                 .build());
             return BaseResponse.builder()
                 .success(false)
