+++ src/main/java/site/cleanfree/be_main/visa/domain/Visa.java
@@ -1,4 +1,4 @@
-package site.cleanfree.be_main.visa;
+package site.cleanfree.be_main.visa.domain;
 
 import jakarta.persistence.Id;
 import lombok.Builder;
@@ -15,10 +15,14 @@ public class Visa extends MongoBaseTimeEntity {
     private String id;
     @Indexed(unique = true)
     private String ip;
+    private String name;
+    private String phoneNumber;
 
     @Builder
-    public Visa(String id, String ip) {
+    public Visa(String id, String ip, String name, String phoneNumber) {
         this.id = id;
         this.ip = ip;
+        this.name = name;
+        this.phoneNumber = phoneNumber;
     }
 }
