+++ src/main/java/site/cleanfree/be_main/visa/vo/VisaRegisterRequestVo.java
@@ -1,0 +1,5 @@
+package site.cleanfree.be_main.visa.vo;
+
+public record VisaRegisterRequestVo(String name, String phoneNumber) {
+
+}
